Make Headstock blocks of different types connect textures
The code in methods 'isIgnoredConnectivitySide' and 'canConnectTexturesToward' will pretend the headstock block styles are the same when they are checked. 

Without this, it will fail the connectivity check and act as if the headstock block beside it is not compatible, even when two adjacent headstock blockstates are identical except for the 'STYLE' value.

To fix this, 'toState', the value temporarily stored from the second compared headstock block is set to have the same 'STYLE' value as 'state', the origin block.

This should work, but I wouldn't like to go through installing Git and build the gradle for two lines of proposed code changed, so I leave this to your discretion to observe, leave, copy over or merge. Thanks for reading all of this, though.